chore: created test files for job applications and job posts endpoints
