+++ server/test/controllers/job_posts/getAllJobPosts.test.js
@@ -1,17 +1,17 @@
 import { expect, vi, describe, it, afterEach, beforeEach } from "vitest";
 
 const mock = require("mock-require");
-const mockGetAllJobPosts = require("../mocks/mockGetAllActiveJobPosts");
+const mockGetAllJobPosts = require("../../mocks/mockGetAllJobPosts");
 
-let getAllActiveJobPostsRequestHandler;
+let getAllJobPostsRequestHandler;
 
 beforeEach(() => {
   // Mock the JobPosting model
   mock("../../src/models/job_posts.model", mockGetAllJobPosts);
 
   // Re-require the handler to apply the mock
-  getAllActiveJobPostsRequestHandler = mock.reRequire(
-    "../../src/controllers/job_posts/getAllActiveJobPosts",
+  getAllJobPostsRequestHandler = mock.reRequire(
+    "../../src/controllers/job_posts/getAllJobPosts",
   );
 });
 
@@ -20,7 +20,7 @@ afterEach(() => {
   mock.stop("../../src/models/job_posts.model");
 });
 
-describe("getAllActiveJobPostsRequestHandler test suite", () => {
+describe("getAllJobPostsRequestHandler test suite", () => {
   const mockReq = {
     method: "GET",
     query: { page: "1", pageSize: "5" },
@@ -42,61 +42,16 @@ describe("getAllActiveJobPostsRequestHandler test suite", () => {
     mockRes.statusCode = 0;
   });
 
+  it("Does not call findOne", async () => {
+    const spy = vi.spyOn(mockGetAllJobPosts, "findOne");
+
+    await getAllJobPostsRequestHandler(mockReq, mockRes);
+    expect(spy).toHaveBeenCalledTimes(0);
+  });
+
   it("Returns 200 on success", async () => {
-    await getAllActiveJobPostsRequestHandler(mockReq, mockRes);
+    await getAllJobPostsRequestHandler(mockReq, mockRes);
     expect(mockRes.statusCode).toBe(200);
-  });
-
-  it("Calls findAndCountAll with the correct state filter", async () => {
-    const spy = vi.spyOn(mockGetAllJobPosts, "findAndCountAll");
-
-    await getAllActiveJobPostsRequestHandler(mockReq, mockRes);
-    expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy).toHaveBeenCalledWith(
-      expect.objectContaining({ where: { state: "Active" } }),
-    );
-  });
-
-  it("Handles pagination correctly", async () => {
-    // Simulate a request for page 2 with pageSize 2
-    mockReq.query = { page: "1", pageSize: "2" };
-
-    await getAllActiveJobPostsRequestHandler(mockReq, mockRes);
-
-    expect(mockRes.statusCode).toBe(200);
-    expect(mockRes.json).toHaveBeenCalledWith({
-      status: "success",
-      message: "All active job posts found successfully",
-      publicJobPosts: {
-        totalPosts: 3, // Only 3 posts with state: "Active"
-        totalPages: 2, // With pageSize of 2, there are 2 total pages
-        currentPage: 1, // Page 1 (0-based indexing)
-        data: [{ id: 3, title: "Web Developer", state: "Active" }],
-      },
-    });
-  });
-
-  it("Handles all active job posts without pagination parameters", async () => {
-    // Simulate a request without pagination parameters
-    mockReq.query = {};
-
-    await getAllActiveJobPostsRequestHandler(mockReq, mockRes);
-
-    expect(mockRes.statusCode).toBe(200);
-    expect(mockRes.json).toHaveBeenCalledWith({
-      status: "success",
-      message: "All active job posts found successfully",
-      publicJobPosts: {
-        totalPosts: 3, // Only 3 posts with state: "Active"
-        totalPages: 1, // No pagination, so all items fit in a single page
-        currentPage: null, // No page provided
-        data: [
-          { id: 1, title: "Software Engineer", state: "Active" },
-          { id: 2, title: "Data Scientist", state: "Active" },
-          { id: 3, title: "Web Developer", state: "Active" },
-        ],
-      },
-    });
   });
 
   it("Handles errors and returns 500", async () => {
@@ -107,11 +62,53 @@ describe("getAllActiveJobPostsRequestHandler test suite", () => {
       },
     );
 
-    await getAllActiveJobPostsRequestHandler(mockReq, mockRes);
+    await getAllJobPostsRequestHandler(mockReq, mockRes);
     expect(mockRes.statusCode).toBe(500);
+  });
+
+  it("Handles pagination correctly", async () => {
+    // Simulate a request for page 2 with pageSize 2
+    mockReq.query = { page: "1", pageSize: "2" };
+
+    await getAllJobPostsRequestHandler(mockReq, mockRes);
+
+    expect(mockRes.statusCode).toBe(200);
     expect(mockRes.json).toHaveBeenCalledWith({
-      status: "error",
-      message: "An unexpected server error occurred.",
+      status: "success",
+      message: "All job posts found successfully",
+      allJobPosts: {
+        totalPosts: 4, // Total posts in the database
+        totalPages: 2, // With pageSize of 2, there are 2 total pages
+        currentPage: 1, // Page 1 (0-based indexing)
+        data: [
+          { id: 3, title: "Web Developer", state: "Active" },
+          { id: 4, title: "DevOps Engineer", state: "Draft" },
+        ],
+      },
+    });
+  });
+
+  it("Handles all mock objects without pagination parameters", async () => {
+    // Simulate a request without pagination parameters
+    mockReq.query = {};
+
+    await getAllJobPostsRequestHandler(mockReq, mockRes);
+
+    expect(mockRes.statusCode).toBe(200);
+    expect(mockRes.json).toHaveBeenCalledWith({
+      status: "success",
+      message: "All job posts found successfully",
+      allJobPosts: {
+        totalPosts: 4, // Total posts in the database
+        totalPages: 1, // No pagination, so all items fit in a single page
+        currentPage: null, // No page provided
+        data: [
+          { id: 1, title: "Software Engineer", state: "Active" },
+          { id: 2, title: "Data Scientist", state: "Active" },
+          { id: 3, title: "Web Developer", state: "Active" },
+          { id: 4, title: "DevOps Engineer", state: "Draft" },
+        ],
+      },
     });
   });
 });
